+++ src/lib/utils/plugins.ts
@@ -3,35 +3,42 @@ import { pathToFileURL } from "url";
 
 import type { Application } from "../application.js";
 import { nicePath } from "./paths.js";
-import { i18n, type TranslatedString } from "#utils";
+import { i18n, type NormalizedPathOrModuleOrFunction, type TranslatedString } from "#utils";
 
 export async function loadPlugins(
     app: Application,
-    plugins: readonly string[],
+    plugins: readonly NormalizedPathOrModuleOrFunction[],
 ) {
     for (const plugin of plugins) {
         const pluginDisplay = getPluginDisplayName(plugin);
 
         try {
-            let instance: any;
-            // Try importing first to avoid warnings about requiring ESM being experimental.
-            // If that fails due to importing a directory, fall back to require.
-            try {
-                // On Windows, we need to ensure this path is a file path.
-                // Or we'll get ERR_UNSUPPORTED_ESM_URL_SCHEME
-                const esmPath = isAbsolute(plugin)
-                    ? pathToFileURL(plugin).toString()
-                    : plugin;
-                instance = await import(esmPath);
-            } catch (error: any) {
-                if (error.code === "ERR_UNSUPPORTED_DIR_IMPORT") {
-                    // eslint-disable-next-line @typescript-eslint/no-require-imports
-                    instance = require(plugin);
-                } else {
-                    throw error;
+            let initFunction: any;
+
+            if (typeof plugin === "function") {
+                initFunction = plugin;
+            } else {
+                let instance: any;
+
+                // Try importing first to avoid warnings about requiring ESM being experimental.
+                // If that fails due to importing a directory, fall back to require.
+                try {
+                    // On Windows, we need to ensure this path is a file path.
+                    // Or we'll get ERR_UNSUPPORTED_ESM_URL_SCHEME
+                    const esmPath = isAbsolute(plugin)
+                        ? pathToFileURL(plugin).toString()
+                        : plugin;
+                    instance = await import(esmPath);
+                } catch (error: any) {
+                    if (error.code === "ERR_UNSUPPORTED_DIR_IMPORT") {
+                        // eslint-disable-next-line @typescript-eslint/no-require-imports
+                        instance = require(plugin);
+                    } else {
+                        throw error;
+                    }
                 }
+                initFunction = instance.load;
             }
-            const initFunction = instance.load;
 
             if (typeof initFunction === "function") {
                 await initFunction(app);
@@ -54,7 +61,11 @@ export async function loadPlugins(
     }
 }
 
-function getPluginDisplayName(plugin: string) {
+function getPluginDisplayName(plugin: NormalizedPathOrModuleOrFunction) {
+    if (typeof plugin === "function") {
+        return plugin.name || "function";
+    }
+
     const path = nicePath(plugin);
     if (path.startsWith("./node_modules/")) {
         return path.substring("./node_modules/".length);